minor guard and test enhancement
git-svn-id: http://svn.codehaus.org/esper/esper/branches/enhancements400@2201 d91eb1b3-1215-0410-9718-d41490442091

+++ esper/src/main/java/com/espertech/esper/pattern/guard/TimerWithinOrMaxCountGuard.java
@@ -81,7 +81,9 @@ public class TimerWithinOrMaxCountGuard implements Guard, ScheduleHandleCallback
     }
 
     private void deactivateTimer() {
-        context.getSchedulingService().remove(scheduleHandle, scheduleSlot);
+        if (scheduleHandle != null) {
+            context.getSchedulingService().remove(scheduleHandle, scheduleSlot);
+        }
         scheduleHandle = null;
         isTimerActive = false;
     }
